requirements.txt
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
pyusb==1.0.2
remotezip==0.9.2
requests==2.22.0
six==1.14.0
soupsieve==2.0
tabulate==0.8.6
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
